refactored ASE datasets for nablaDFT
