+++ config/datamodule/nablaDFT_ase.yaml
@@ -4,7 +4,6 @@ _target_: nablaDFT.dataset.ASENablaDFT
 split: ${job_type}
 datapath: ./datasets/nablaDFT/train/raw
 dataset_name: ${dataset_name}
-data_workdir: ./spk_logs
 train_ratio: 0.9
 val_ratio: 0.1
 batch_size: 32
