+++ nablaDFT/dataset/nablaDFT_dataset.py
@@ -1,27 +1,32 @@
 """Module defines Pytorch Lightning DataModule interfaces for various NablaDFT datasets"""
-
+from typing import List, Dict, Optional, Union
 import json
 import os
-from typing import Optional
 from urllib import request as request
 
 from tqdm import tqdm
 import numpy as np
-import torch
 from ase.db import connect
+import torch
 from pytorch_lightning import LightningDataModule
 from torch.utils.data import random_split
 from torch_geometric.loader import DataLoader
-from schnetpack.data import AtomsDataFormat, load_dataset
+from schnetpack.data import (
+    AtomsDataFormat,
+    load_dataset,
+    AtomsLoader,
+    SplittingStrategy,
+    AtomsDataModule,
+)
 
 import nablaDFT
 from nablaDFT.utils import tqdm_download_hook, get_file_size
-from .atoms_datamodule import AtomsDataModule
 from .pyg_datasets import PyGNablaDFT, PyGHamiltonianNablaDFT
 
 
 class ASENablaDFT(AtomsDataModule):
     """PytorchLightning interface for nablaDFT ASE datasets.
+    Overrides schnetpack.AtomsDataModule.
 
     Args:
         split (str): type of split, must be one of ['train', 'test', 'predict'].
@@ -33,36 +38,72 @@ class ASENablaDFT(AtomsDataModule):
         train_transforms (Callable): data transform, called for every sample in training dataset.
         val_transforms (Callable): data transform, called for every sample in validation dataset.
         test_transforms (Callable): data transform, called for every sample in test dataset.
+
+    Other args description could be found in SchNetPack docs:
+    https://schnetpack.readthedocs.io/en/latest/api/generated/data.AtomsDataModule.html#data.AtomsDataModule
     """
+
+    format = AtomsDataFormat.ASE
 
     def __init__(
         self,
+        dataset_name: str,
         split: str,
-        dataset_name: str = "dataset_train_tiny",
-        datapath: str = "database",
-        data_workdir: Optional[str] = "logs",
-        batch_size: int = 2000,
+        datapath: str,
+        batch_size: int,
         train_ratio: float = 0.9,
         val_ratio: float = 0.1,
         test_ratio: float = 0.0,
-        train_transforms=None,
-        val_transforms=None,
-        test_transforms=None,
-        format: Optional[AtomsDataFormat] = AtomsDataFormat.ASE,
-        **kwargs,
+        data_workdir: Optional[str] = None,
+        num_train: Union[int, float] = None,
+        num_val: Union[int, float] = None,
+        num_test: Optional[Union[int, float]] = None,
+        split_file: Optional[str] = "split.npz",
+        load_properties: Optional[List[str]] = None,
+        val_batch_size: Optional[int] = None,
+        test_batch_size: Optional[int] = None,
+        transforms: Optional[List[torch.nn.Module]] = None,
+        train_transforms: Optional[List[torch.nn.Module]] = None,
+        val_transforms: Optional[List[torch.nn.Module]] = None,
+        test_transforms: Optional[List[torch.nn.Module]] = None,
+        num_workers: int = 8,
+        num_val_workers: Optional[int] = None,
+        num_test_workers: Optional[int] = None,
+        property_units: Optional[Dict[str, str]] = None,
+        distance_unit: Optional[str] = None,
+        cleanup_workdir_stage: Optional[str] = "test",
+        splitting: Optional[SplittingStrategy] = None,
+        pin_memory: Optional[bool] = False,
     ):
         """"""
         super().__init__(
-            split=split,
             datapath=datapath,
-            data_workdir=data_workdir,
             batch_size=batch_size,
+            num_train=num_train,
+            num_val=num_val,
+            num_test=num_test,
+            split_file=split_file,
+            format=self.format,
+            load_properties=load_properties,
+            val_batch_size=val_batch_size,
+            test_batch_size=test_batch_size,
+            transforms=transforms,
             train_transforms=train_transforms,
             val_transforms=val_transforms,
             test_transforms=test_transforms,
-            format=format,
-            **kwargs,
+            num_workers=num_workers,
+            num_val_workers=num_val_workers,
+            num_test_workers=num_test_workers,
+            property_units=property_units,
+            distance_unit=distance_unit,
+            data_workdir=data_workdir,
+            cleanup_workdir_stage=cleanup_workdir_stage,
+            splitting=splitting,
+            pin_memory=pin_memory,
         )
+        self.split = split
+        self._predict_dataset = None
+        self._predict_dataloader = None
         self.dataset_name = dataset_name
         self.train_ratio = train_ratio
         self.val_ratio = val_ratio
@@ -108,7 +149,52 @@ class ASENablaDFT(AtomsDataModule):
             if not self.num_train and not self.num_val:
                 self.num_val = -1
                 self.num_train = -1
-        self.dataset = load_dataset(self.datapath, self.format)
+
+    def setup(self, stage: Optional[str] = None):
+        """Overrides method from original AtomsDataModule class
+
+        Args:
+            stage (str): trainer stage, must be one of ['fit', 'test', 'predict']
+        """
+        # check whether data needs to be copied
+        if self.data_workdir is None:
+            datapath = self.datapath
+        else:
+            datapath = self._copy_to_workdir()
+        # (re)load datasets
+        if self.dataset is None:
+            self.dataset = load_dataset(
+                datapath,
+                self.format,
+                property_units=self.property_units,
+                distance_unit=self.distance_unit,
+                load_properties=self.load_properties,
+            )
+
+            # load and generate partitions if needed
+            if self.train_idx is None:
+                self._load_partitions()
+
+            # partition dataset
+            self._train_dataset = self.dataset.subset(self.train_idx)
+            self._val_dataset = self.dataset.subset(self.val_idx)
+            if self.split == "preidct":
+                self._predict_dataset = self.dataset.subset(self.test_idx)
+            else:
+                self._test_dataset = self.dataset.subset(self.test_idx)
+            self._setup_transforms()
+
+    def predict_dataloader(self) -> AtomsLoader:
+        """Describes predict dataloader, used for prediction task"""
+        if self._predict_dataloader is None:
+            self._predict_dataloader = AtomsLoader(
+                self._predict_dataset,
+                batch_size=self.test_batch_size,
+                num_workers=self.num_test_workers,
+                pin_memory=self._pin_memory,
+                shuffle=False,
+            )
+        return self._predict_dataloader
 
 
 class PyGDataModule(LightningDataModule):
